+++ lib/rails_presenter.rb
@@ -1,4 +1,6 @@
-require "rails_presenter/engine"
+require 'rails_presenter/engine'
+require 'rails_presenter/presenter_helper'
+require 'rails_presenter/base'
 
 module RailsPresenter
 end
